// packages/oraiswap-v3/src/wasm/oraiswap_v3_wasm.d.ts
export interface FeeTier {
    fee: Percentage;
    tick_spacing: number;
}

export interface Pool {
    liquidity: Liquidity;
    sqrt_price: SqrtPrice;
    current_tick_index: number;
    fee_growth_global_x: FeeGrowth;
    fee_growth_global_y: FeeGrowth;
    fee_protocol_token_x: TokenAmount;
    fee_protocol_token_y: TokenAmount;
    start_timestamp: number;
    last_timestamp: number;
    fee_receiver: string;
}

export interface PoolKey {
    token_x: string;
    token_y: string;
    fee_tier: FeeTier;
}
// ...
export interface LiquidityTick {
    index: number;
    liquidity_change: Liquidity;
    sign: boolean;
}

export type LiquidityTickVec = LiquidityTick[];

export type FeeGrowth = bigint;

export type FixedPoint = bigint;

export type Liquidity = bigint;

export type Percentage = number;